Rename ClineAPI to RooCodeAPI and improve types
