+++ src/activate/index.ts
@@ -1,3 +1,4 @@
 export { handleUri } from "./handleUri"
 export { registerCommands } from "./registerCommands"
 export { registerCodeActions } from "./registerCodeActions"
+export { createRooCodeAPI } from "./createRooCodeAPI"
